API

Contents

API

General

Bags

Data nodes

Index

Mill.AbstractBagNode

Mill.AbstractBags

Mill.AbstractNode

Mill.AbstractProductNode

Mill.AlignedBags

Mill.AlignedBags

Mill.AlignedBags

Mill.AlignedBags

Mill.ScatteredBags

Mill.ScatteredBags

Mill.ScatteredBags

Mill.adjustbags

Mill.bagcount

Mill.bagcount!

Mill.bags

Mill.data

Mill.emptyismissing

Mill.emptyismissing!

Mill.length2bags

Mill.metadata

Mill.remapbags

Mill.string_end_code

Mill.string_end_code!

Mill.string_start_code

Mill.string_start_code!

General

Mill.emptyismissing — Function

Mill.emptyismissing()

Get the current value of the emptyismissing parameter.

See also: Mill.emptyismissing!, Empty bags (docs).

Mill.emptyismissing! — Function

Mill.emptyismissing!(::Bool)

Set the new value to the emptyismissing parameter.

See also: Mill.emptyismissing, Empty bags (docs).

Mill.bagcount — Function

Mill.bagcount()

Get the current value of the bagcount parameter.

See also: Mill.bagcount!, Bag count (docs).

Mill.bagcount! — Function

Mill.bagcount!(Bool)

Set the new value to the bagcount parameter.

See also: Mill.bagcount, Bag count (docs).

Mill.string_start_code — Function

Mill.string_start_code()

Get the current value of the string_start_code parameter used as a code point of the abstract string-start character. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding.

See also: Mill.string_start_code!, Mill.string_end_code, Mill.string_end_code!, Strings (docs).

Mill.string_start_code! — Function

Mill.string_start_code!(c::UInt8)

Set the new value to the string_start_code parameter used as a code point of the abstract string-start character to c. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding.

See also: Mill.string_start_code, Mill.string_end_code, Mill.string_end_code!, Strings (docs).

Mill.string_end_code — Function

Mill.string_end_code()

Get the current value of the string_end_code parameter used as a code point of the abstract string-end character. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding.

See also: Mill.string_end_code!, Mill.string_start_code, Mill.string_start_code!, Strings (docs).

Mill.string_end_code! — Function

Mill.string_end_code!(c::UInt8)

Set the new value to the string_end_code parameter used as a code point of the abstract string-end character to c. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding.

See also: Mill.string_end_code, Mill.string_start_code, Mill.string_start_code!, Strings (docs).

Bags

Mill.AbstractBags — Type

AbstractBags{T}

Supertype for structures storing indices of type T of bags' instances in BagNodes.

Mill.AlignedBags — Type

AlignedBags{T <: Integer} <: AbstractBags{T}

AlignedBags struct stores indices of bags' instances in one or more UnitRange{T}s. This is only possible if instances of every bag are stored in one contiguous block.

See also: ScatteredBags.

Mill.AlignedBags — Method

AlignedBags()

Construct new AlignedBags struct containing no bags.

Examples

julia> AlignedBags()
AlignedBags{Int64}(UnitRange{Int64}[])

Mill.AlignedBags — Method

AlignedBags(bags::UnitRange{<:Integer}...)

Construct new AlignedBags struct from bags in arguments.

Examples

julia> AlignedBags(1:3, 4:8)
AlignedBags{Int64}(UnitRange{Int64}[1:3, 4:8])

Mill.AlignedBags — Method

AlignedBags(k::Vector{<:Integer})

Construct new AlignedBags struct from Vectork specifying the index of the bag each instance belongs to. Throws ArgumentError if this is not possible.

Examples

julia> AlignedBags([2, 2, 1, 1, 1, 3])
AlignedBags{Int64}(UnitRange{Int64}[1:2, 3:5, 6:6])

Mill.ScatteredBags — Type

ScatteredBags{T <: Integer} <: AbstractBags{T}

ScatteredBags struct stores indices of bags' instances that are not necessarily contiguous.

See also: AlignedBags.

Mill.ScatteredBags — Method

ScatteredBags()

Construct new ScatteredBags struct containing no bags.

Examples

julia> ScatteredBags()
ScatteredBags{Int64}(Array{Int64,1}[])

Mill.ScatteredBags — Method

ScatteredBags(k::Vector{<:Integer})

Construct new ScatteredBags struct from Vectork specifying the index of the bag each instance belongs to.

Examples

julia> ScatteredBags([2, 2, 1, 1, 1, 3])
ScatteredBags{Int64}([[3, 4, 5], [1, 2], [6]])

Mill.length2bags — Function

length2bags(ls::Vector{<:Integer})

Convert lengths of bags given in ls to AlignedBags with contiguous blocks.

Examples

julia> length2bags([1, 3, 2])
AlignedBags{Int64}(UnitRange{Int64}[1:1, 2:4, 5:6])

See also: AlignedBags.

Mill.bags — Function

bags(k::Vector{<:Integer})
bags(k::Vector{T}) where T <: UnitRange{<:Integer}
bags(b::AbstractBags)

Construct an AbstractBags structure that is most suitable for the input (AlignedBags if possible, ScatteredBags otherwise).

Examples

julia> bags([2, 2, 3, 1])
AlignedBags{Int64}(UnitRange{Int64}[1:2, 3:3, 4:4])

julia> bags([2, 3, 1, 2])
ScatteredBags{Int64}([[3], [1, 4], [2]])

julia> bags([1:3, 4:5])
AlignedBags{Int64}(UnitRange{Int64}[1:3, 4:5])

julia> bags(ScatteredBags())
ScatteredBags{Int64}(Array{Int64,1}[])

See also: AlignedBags, ScatteredBags.

Mill.remapbags — Function

remapbags(b::AbstractBags, idcs::VecOrRange{<:Integer}) -> (rb, I)

Select a subset of bags in b corresponding to indices idcs and remap instance indices appropriately. Return new bags rb as well as a Vector of remapped instances I.

Examples

julia> remapbags(AlignedBags([1:1, 2:3, 4:5]), [1, 3])
(AlignedBags{Int64}(UnitRange{Int64}[1:1, 2:3]), [1, 4, 5])

julia> remapbags(ScatteredBags([[1,3], [2], Int[]]), [2])
(ScatteredBags{Int64}([[1]]), [2])

Mill.adjustbags — Function

adjustbags(b::AlignedBags, mask::AbstractVector{Bool})

Remove indices of instances brom bags b and remap the remaining instances accordingly.

Examples

julia> adjustbags(AlignedBags([1:2, 0:-1, 3:4]), [false, false, true, true])
AlignedBags{Int64}(UnitRange{Int64}[0:-1, 0:-1, 1:2])

Data nodes

Mill.AbstractNode — Type

AbstractNode

Supertype for any structure representing a data node.

Mill.AbstractProductNode — Type

AbstractProductNode <: AbstractNode

Supertype for any structure representing a data node implementing a Cartesian product of data in subtrees.

Mill.AbstractBagNode — Type

AbstractBagNode <: AbstractNode

Supertype for any structure representing a data node that implements a multi-instance learning problem.

Mill.data — Function

Mill.data(n::AbstractNode)

Return data stored in node n.

Examples

julia> Mill.data(ArrayNode([1 2; 3 4], "metadata"))
2×2 Array{Int64,2}:
 1  2
 3  4

julia> Mill.data(BagNode(ArrayNode([1 2; 3 4]), bags([1:3, 4:4]), "metadata"))
2×2 ArrayNode{Array{Int64,2},Nothing}:
 1  2
 3  4

See also: Mill.metadata

Mill.metadata — Function

Mill.metadata(n::AbstractNode)

Return metadata stored in node n.

Examples

julia> Mill.metadata(ArrayNode([1 2; 3 4], "metadata"))
"metadata"

julia> Mill.metadata(BagNode(ArrayNode([1 2; 3 4]), bags([1:3, 4:4]), "metadata"))
"metadata"

See also: Mill.data