General

Mill.emptyismissing — Function

Returns the current value of the emptyismissing parameter. See the Empty bags section of the docs for more information.

Mill.emptyismissing! — Function

Sets the new value of the emptyismissing parameter. See the Empty bags section of the docs for more information.

Mill.bagcount — Function

Returns the current value of the bagcount parameter. See the Bag count section of the docs for more information.

Mill.bagcount! — Function

Sets the new value of the bagcount parameter. See the Bag count section of the docs for more information.

Mill.string_start_code — Function

Mill.string_start_code()

Returns the current value of the string_start_code parameter used as a code point of the abstract string-start character. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding. See the Strings section of the docs for more information.

Mill.string_start_code! — Function

Mill.string_start_code!(c::UInt8)

Sets the new value of the string_start_code parameter used as a code point of the abstract string-start character to c. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding. See the Strings section of the docs for more information.

Mill.string_end_code — Function

Mill.string_end_code()

Returns the current value of the string_end_code parameter used as a code point of the abstract string-end character. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding. See the Strings section of the docs for more information.

Mill.string_end_code! — Function

Mill.string_end_code!(c::UInt8)

Sets the new value of the string_end_code parameter used as a code point of the abstract string-end character to c. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding. See the Strings section of the docs for more information.