General

Index

API

Mill.emptyismissing — Function

Mill.emptyismissing()

Get the current value of the emptyismissing parameter.

See also: Mill.emptyismissing!.

Mill.emptyismissing! — Function

Mill.emptyismissing!(::Bool; persist=false)

Set the new value to the emptyismissing parameter.

Set persist=true to persist this setting between sessions.

See also: Mill.emptyismissing.

Mill.string_start_code — Function

Mill.string_start_code()

Get the current value of the string_start_code parameter used as a code point of the abstract string-start character. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding.

See also: Mill.string_start_code!, Mill.string_end_code, Mill.string_end_code!.

Mill.string_start_code! — Function

Mill.string_start_code!(c::Integer; persist=false)

Set the new value to the string_start_code parameter used as a code point of the abstract string-start character to c. The default value of the parameter is 0x02, which corresponds to the STX character in ASCII encoding.

c should fit into UInt8.

Set persist=true to persist this setting between sessions.

See also: Mill.string_start_code, Mill.string_end_code, Mill.string_end_code!.

Mill.string_end_code — Function

Mill.string_end_code()

Get the current value of the string_end_code parameter used as a code point of the abstract string-end character. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding.

See also: Mill.string_end_code!, Mill.string_start_code, Mill.string_start_code!.

Mill.string_end_code! — Function

Mill.string_end_code!(c::Integer; persist=false)

Set the new value to the string_end_code parameter used as a code point of the abstract string-end character to c. The default value of the parameter is 0x03, which corresponds to the ETX character in ASCII encoding.

c should fit into UInt8.

Set persist=true to persist this setting between sessions.

See also: Mill.string_end_code, Mill.string_start_code, Mill.string_start_code!.